Massive refactoring on the pipelines to clean up some ugliness

+++ common/build-common.sh
@@ -9,28 +9,37 @@ if [ -n "${BASH_SOURCE[1]}" ]; then
         export DOCKER_REPO="entando/$ENTANDO_IMAGE"
         export IMAGE_NAME="$DOCKER_REPO:$VERSION"
         export DOCKERFILE_PATH="$DOCKER_BUILD_DIR/Dockerfile"
-        $DOCKER_BUILD_DIR/hooks/build
+        $DOCKER_BUILD_DIR/hooks/build docker &> $DOCKER_BUILD_DIR/docker-build.log || exit 1
         if [ $? -eq 0 ]; then
-            #Now push to locally forwarded port (easiest way to get access to Openshift registry)
-            docker login -u $(oc whoami) -p $(oc whoami -t) $OPENSHIFT_REGISTRY
-            export DOCKER_REPO="$OPENSHIFT_REGISTRY/$DOCKER_REPO"
-
-            echo "Pushing $DOCKER_REPO:$DOCKER_TAG to local registry"
-            docker tag $IMAGE_NAME "$DOCKER_REPO:$DOCKER_TAG"
-            docker push "$DOCKER_REPO:$DOCKER_TAG"
-            echo "Removing the local tag $DOCKER_REPO:$DOCKER_TAG"
-            docker rmi "$DOCKER_REPO:$DOCKER_TAG"
-
-            echo "Optionally tagging and pushing $DOCKER_REPO:latest"
-            $DOCKER_BUILD_DIR/hooks/post_push
-
+            echo "Docker build successful"
+            if [ "$PUSH_TO_DOCKER_HUB" = true ]; then
+                docker push "$DOCKER_REPO:$DOCKER_TAG"
+            fi
             LATEST_VERSION=$(cat $(dirname $(realpath ${BASH_SOURCE[0]}))/hooks/VERSION_TO_TAG_AS_LATEST)
-            #If we are building the latest release branch, tag as latest
+            #If we are building the latest release branch, tag as latest. This would usually be done in post_push but when running on
+            # our own infrastructure we don't have a post_push callback
             if [ "$VERSION" = "$LATEST_VERSION" ]; then
                 echo "Tagging entando/$ENTANDO_IMAGE:latest"
                 docker tag "entando/$ENTANDO_IMAGE:$VERSION" "entando/$ENTANDO_IMAGE:latest"
-                echo "Removing useless $DOCKER_REPO:$DOCKER_TAG tag"
-                docker rmi "$DOCKER_REPO:latest"
+                if [ "$PUSH_TO_DOCKER_HUB" = true ]; then
+                    docker push "entando/$ENTANDO_IMAGE:latest"
+                fi
+            fi
+            if [ -n "$OPENSHIFT_REGISTRY" ]; then
+                #Now push to locally forwarded port (easiest way to get access to Openshift registry)
+                docker login -u $(oc whoami) -p $(oc whoami -t) $OPENSHIFT_REGISTRY
+                export DOCKER_REPO="$OPENSHIFT_REGISTRY/$DOCKER_REPO"
+
+                echo "Pushing $DOCKER_REPO:$DOCKER_TAG to local registry"
+                docker tag $IMAGE_NAME "$DOCKER_REPO:$DOCKER_TAG"
+                docker push "$DOCKER_REPO:$DOCKER_TAG"
+                echo "Removing the local tag $DOCKER_REPO:$DOCKER_TAG"
+                docker rmi "$DOCKER_REPO:$DOCKER_TAG"
+
+                echo "Optionally tagging and pushing $DOCKER_REPO:latest"
+                $DOCKER_BUILD_DIR/hooks/post_push # this is more to just test that post_push works, testing it against Openshift's docker registry
+                echo "Removing local tag $DOCKER_REPO:latest if present"
+                docker rmi "$DOCKER_REPO:latest" &> /dev/null
             fi
         else
             echo "Docker build failed"
