+++ common/s2i/init-derby-from-war.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 #NB!!! This file is copied from common/s2i in the Docker build hook. Only modify the original file!
 #Rebuilds a Derby database for embedded use IF and only if the database driver is derby and PREPARE_DATA is true
-if [ "$PORTDB_DRIVER" = "derby" ]; then
+if [ "$PORTDB_DRIVER" = "derby" ] && [ "$SERVDB_DRIVER" = "derby" ]; then
   if [ "$PREPARE_DATA" = "true" ]; then
     rm -Rf /entando-data-templates/databases/
     if [ -d /entando-data/databases ]; then
@@ -20,7 +20,6 @@ if [ "$PORTDB_DRIVER" = "derby" ]; then
       chown -Rf $USERID_TO_USE:0 /entando-data-templates/databases/
       rm -Rf /entando-data/protected
       rm -Rf /entando-data/resources
-      echo $(date +%s) > /entando-data-templates/build_id
     else
       echo "Derby Database Build failed"
       exit 1
@@ -29,4 +28,6 @@ if [ "$PORTDB_DRIVER" = "derby" ]; then
     echo "Derby Database not built because PREPARE_DATA=false"
   fi
 fi
+#Update the build_id whether there was a Derby rebuild or not. This is to be used in prepare-date.sh
+echo $(date +%s) > /entando-data-templates/build_id
 exit 0
